+++ components/com_kinoarhiv/views/movie/view.html.php
@@ -44,7 +44,7 @@ class KinoarhivViewMovie extends JViewLegacy
 	 *
 	 * @param   string  $tpl  The name of the template file to parse; automatically searches through the template paths.
 	 *
-	 * @return  mixed
+	 * @return  void
 	 *
 	 * @since  3.0
 	 */